refactor data cleaning
